task_3_watermark_detection/requirements.txt
torch
torchvision
pandas
requests
scikit-learn
numpy
pillow